+++ crud_operations.py
@@ -1,0 +1,15 @@
+def add_student():
+    # Logik för att lägga till student
+    pass
+
+def view_students():
+    # Logik för att visa studenter
+    pass
+
+def update_student():
+    # Logik för att uppdatera en student
+    pass
+
+def delete_student():
+    # Logik för att ta bort en student
+    pass
